[Rename] :  donation 폴더 분리
